Add TV Remote Input API

+++ ism/extras/efl_panel/motion_input.cpp
@@ -30,7 +30,7 @@ void hpfitler(const _Tp* Y, _Tp* Trend, int numObs, double smoothing)
     }
     else
     {
-        float e[]={smoothing,-4*smoothing,(1+6*smoothing),-4*smoothing,smoothing};
+        float e[]={(float)smoothing,-4*(float)smoothing,(1+6*(float)smoothing),-4*(float)smoothing,(float)smoothing};
         for (int i=0;i<numObs;i++)
         {
             cvSetReal2D(A,i,i,e[2]);
@@ -66,6 +66,7 @@ void hpfitler(const _Tp* Y, _Tp* Trend, int numObs, double smoothing)
     cvReleaseMat(&b);
     cvReleaseMat(&A);
 }
+
 void hpfitler(std::deque<Point2Df>& pY, std::deque<Point2Df>& pTrend, double smoothing)
 {
     if (pTrend.empty())
@@ -91,6 +92,7 @@ void hpfitler(std::deque<Point2Df>& pY, std::deque<Point2Df>& pTrend, double smo
     delete[] Y;
     delete[] Trend;
 }
+
 void hpfitler(std::deque<Point3Df>& pY, std::deque<Point3Df>& pTrend, double smoothing)
 {
     if (pTrend.empty())
@@ -151,7 +153,7 @@ bool Motion_Input::filter_raw_sensor_data(Point3Df *origin_AccData, Point3Df *or
     _GyrFilterBuffer.push_back(GyrData);
     _AccFilterBuffer.push_back(AccData);
 
-    if (_GyrFilterBuffer.size() > _filterNumber)
+    if (_GyrFilterBuffer.size() > (unsigned int)_filterNumber)
     {
         _GyrFilterBuffer.pop_front();
         std::deque<Point3Df> filter_points(_GyrFilterBuffer.size());
@@ -172,7 +174,7 @@ bool Motion_Input::filter_raw_sensor_data(Point3Df *origin_AccData, Point3Df *or
         GyrData.z = zoom_m * (GyrData.z * (1 - rate) + filter_points[_filterNumber - 1].z * rate);
     }
 
-    if (_AccFilterBuffer.size() > _filterNumber)
+    if (_AccFilterBuffer.size() > (unsigned int)_filterNumber)
     {
         _AccFilterBuffer.pop_front();
         std::deque<Point3Df> filter_points(_AccFilterBuffer.size());
